requirements.txt
pandas
scikit-learn
joblib
groq
tqdm
python-dotenv
gradio